feat(buyer/products): improve search products bar

+++ database/migrations/2022_01_03_100913_create_products_table.php
@@ -12,7 +12,7 @@ class CreateProductsTable extends Migration
             $table->id();
             $table->string('name');
             $table->text('description');
-            $table->bigInteger('price');
+            $table->integer('price');
             $table->bigInteger('stock')->nullable();
             $table->foreignId('category_id')
                 ->nullable()
